standarize styles to match figma H4
