+++ src/app/auth/styles.tsx
@@ -1,23 +1,25 @@
 import styled from 'styled-components';
+import { Open_Sans as OpenSans } from 'next/font/google';
 import COLORS from '@/styles/colors';
-import { H4 } from '@/styles/text';
 
-export const OuterDiv = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 100vw;
-  height: 100vh;
-`;
+const openSans = OpenSans({ subsets: ['latin'] });
 
 export const FormDiv = styled.div`
   display: flex;
   flex-direction: column;
   gap: 2.5rem; // 40 px
   padding: 3.25rem; //52px = 13/4
-  border-radius: 0.625rem;
-  border: 2px solid ${COLORS.blueMid};
-  width: 580px;
+  border-radius: 10px;
+  border: 2px solid ${COLORS.activeBlue};
+`;
+
+export const ForgotPassword = styled.p`
+  color: ${COLORS.midDarkGrey};
+  ${openSans.style}
+  font-size: 0.875rem; // 14px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: normal;
 `;
 
 export const QuestionsDiv = styled.div`
@@ -26,13 +28,8 @@ export const QuestionsDiv = styled.div`
   gap: 1.2rem; // 20 px
 `;
 
-export const SpacerDiv = styled.div`
+export const SixteenDiv = styled.div`
   display: flex;
   flex-direction: column;
   gap: 1rem;
-`;
-
-export const H4Centered = styled(H4)`
-  color: ${COLORS.greyDark};
-  text-align: center;
 `;
