+++ src/components/TextInput/styles.tsx
@@ -1,5 +1,6 @@
 import styled from 'styled-components';
 import { Open_Sans as OpenSans } from 'next/font/google';
+// import { H4 } from '@/styles/text';
 import COLORS from '../../styles/colors';
 
 const openSans = OpenSans({ subsets: ['latin'] });
